minor ui uplift
