+++ .github/workflows/core-dependency-update.yml
@@ -93,5 +93,5 @@ jobs:
             **Status**: ${{ job.status == 'success' && '✅ Success' || '❌ Failed' }}
             **Branch**: `main`
             **Commit**: ```${{ github.sha }}```
-            **Author**: ${{ github.actor }}
+            **Author**: ${{ github.event.head_commit.author.username || github.event.head_commit.author.name || github.event.sender.login || github.actor }}
             **[View Workflow Run](${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }})**
