+++ ui/app/config/views/mcp-client-form.tsx
