+++ ui/app/governance/views/customer-dialog.tsx
@@ -1,22 +1,22 @@
 'use client'
 
+import FormFooter from '@/components/form-footer'
+import { Badge } from '@/components/ui/badge'
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
 import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle } from '@/components/ui/dialog'
 import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
-import { apiService } from '@/lib/api'
-import { Customer, CreateCustomerRequest, UpdateCustomerRequest } from '@/lib/types/governance'
-import { User, DollarSign } from 'lucide-react'
-import { useState, useEffect, useMemo } from 'react'
-import { toast } from 'sonner'
 import NumberAndSelect from '@/components/ui/number-and-select'
+import { apiService } from '@/lib/api'
 import { resetDurationOptions } from '@/lib/constants/governance'
-import { Badge } from '@/components/ui/badge'
-import { Validator } from '@/lib/utils/validation'
-import isEqual from 'lodash.isequal'
-import FormFooter from '../form-footer'
-import { formatDistanceToNow } from 'date-fns'
+import { CreateCustomerRequest, Customer, UpdateCustomerRequest } from '@/lib/types/governance'
 import { formatCurrency } from '@/lib/utils/governance'
+import { Validator } from '@/lib/utils/validation'
+import { formatDistanceToNow } from 'date-fns'
+import isEqual from 'lodash.isequal'
+import { DollarSign } from 'lucide-react'
+import { useEffect, useMemo, useState } from 'react'
+import { toast } from 'sonner'
 
 interface CustomerDialogProps {
   customer?: Customer | null
