+++ ui/app/governance/views/team-dialog.tsx
@@ -1,23 +1,22 @@
 'use client'
 
+import FormFooter from '@/components/form-footer'
+import { Badge } from '@/components/ui/badge'
 import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle } from '@/components/ui/dialog'
 import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
+import NumberAndSelect from '@/components/ui/number-and-select'
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select'
 import { apiService } from '@/lib/api'
-import { Team, Customer, CreateTeamRequest, UpdateTeamRequest } from '@/lib/types/governance'
-import { User } from 'lucide-react'
-import { useState, useEffect, useMemo } from 'react'
-import { toast } from 'sonner'
-import NumberAndSelect from '@/components/ui/number-and-select'
 import { resetDurationOptions } from '@/lib/constants/governance'
-import { Badge } from '@/components/ui/badge'
-import { Validator } from '@/lib/utils/validation'
-import isEqual from 'lodash.isequal'
-import FormFooter from '../form-footer'
-import { formatDistanceToNow } from 'date-fns'
-import { parseResetPeriod } from '@/lib/utils/governance'
+import { CreateTeamRequest, Customer, Team, UpdateTeamRequest } from '@/lib/types/governance'
 import { formatCurrency } from '@/lib/utils/governance'
+import { Validator } from '@/lib/utils/validation'
+import { formatDistanceToNow } from 'date-fns'
+import isEqual from 'lodash.isequal'
+import { User } from 'lucide-react'
+import { useEffect, useMemo, useState } from 'react'
+import { toast } from 'sonner'
 
 interface TeamDialogProps {
   team?: Team | null
