+++ ui/app/governance/views/virtual-key-detail-dialog.tsx
