+++ ui/app/logs/views/audio-player.tsx
