+++ ui/app/logs/views/code-editor.css
