+++ ui/app/logs/views/code-editor.tsx
