+++ ui/app/logs/views/log-entry-details-view.tsx
