+++ ui/app/logs/views/log-message-view.tsx
