+++ ui/app/logs/views/transcription-view.tsx
