+++ ui/components/ui/skeleton.tsx
@@ -1,7 +1,7 @@
 import { cn } from "@/lib/utils";
 
 function Skeleton({ className, ...props }: React.ComponentProps<"div">) {
-	return <div data-slot="skeleton" className={cn("bg-accent animate-pulse rounded-md", className)} {...props} />;
+	return <div data-slot="skeleton" className={cn("bg-accent animate-pulse rounded-sm", className)} {...props} />;
 }
 
 export { Skeleton };
