+++ ui/components/ui/tag-input.tsx
@@ -1,10 +1,10 @@
 "use client";
 
-import React from "react";
-import { Input } from "@/components/ui/input";
 import { Badge } from "@/components/ui/badge";
-import { X } from "lucide-react";
+import { Input } from "@/components/ui/input";
 import { cn } from "@/lib/utils";
+import { X } from "lucide-react";
+import React from "react";
 
 type OmittedInputProps = Omit<React.InputHTMLAttributes<HTMLInputElement>, "value" | "onChange">;
 
@@ -46,7 +46,7 @@ export const TagInput = React.forwardRef<HTMLInputElement, TagInputProps>(({ cla
 	};
 
 	return (
-		<div className={cn("border-input flex flex-wrap items-center gap-2 rounded-md border p-2", className)}>
+		<div className={cn("border-input flex flex-wrap items-center gap-2 rounded-sm border p-2", className)}>
 			{value.map((tag) => (
 				<Badge key={tag} variant="secondary" className="flex items-center gap-1">
 					{tag}
